+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/remove_service/remove_service.go
@@ -77,7 +77,7 @@ func (instruction *RemoveServiceInstruction) String() string {
 
 func (instruction *RemoveServiceInstruction) ValidateAndUpdateEnvironment(environment *startosis_validator.ValidatorEnvironment) error {
 	if !environment.DoesServiceIdExist(instruction.serviceId) {
-		return stacktrace.NewError("There was an error validating remove service as service ID '%v' doesn't exist", instruction.serviceId)
+		return startosis_errors.NewValidationError("There was an error validating remove service as service ID '%v' doesn't exist", instruction.serviceId)
 	}
 	environment.RemoveServiceId(instruction.serviceId)
 	return nil