+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/wait/wait.go
@@ -82,7 +82,7 @@ func (instruction *WaitInstruction) String() string {
 
 func (instruction *WaitInstruction) ValidateAndUpdateEnvironment(environment *startosis_validator.ValidatorEnvironment) error {
 	if !environment.DoesServiceIdExist(instruction.serviceId) {
-		return stacktrace.NewError("There was an error validating exec with service ID '%v' that does not exist", instruction.serviceId)
+		return startosis_errors.NewValidationError("There was an error validating exec with service ID '%v' that does not exist", instruction.serviceId)
 	}
 	// TODO(victor.colombo): Add fact validation
 	return nil